scalability framework: split and move workloads file
