+++ misc/python/materialize/scalability/workloads/dml_dql_workloads.py
@@ -8,32 +8,18 @@
 # by the Apache License, Version 2.0.
 
 
-from materialize.scalability.operation import Operation, OperationChainWithDataExchange
-from materialize.scalability.operation_data import OperationData
+from materialize.scalability.operation import Operation
 from materialize.scalability.operations import (
-    Connect,
-    CreateMvOnTableX,
-    CreateTableX,
-    Disconnect,
-    DropTableX,
     InsertDefaultValues,
-    PopulateTableX,
     SelectCount,
     SelectCountInMv,
     SelectLimit,
     SelectOne,
     SelectStar,
-    SelectStarFromMvOnTableX,
-    SelectStarFromTableX,
     SelectUnionAll,
     Update,
 )
-from materialize.scalability.workload import (
-    WorkloadWithContext,
-)
 from materialize.scalability.workload_markers import (
-    ConnectionWorkload,
-    DdlWorkload,
     DmlDqlWorkload,
 )
 
@@ -81,43 +67,3 @@ class InsertAndSelectLimitWorkload(DmlDqlWorkload):
 class UpdateWorkload(DmlDqlWorkload):
     def operations(self) -> list["Operation"]:
         return [Update()]
-
-
-class EstablishConnectionWorkload(WorkloadWithContext, ConnectionWorkload):
-    def amend_data_before_execution(self, data: OperationData) -> None:
-        data.push("endpoint", self.endpoint)
-        data.push("schema", self.schema)
-        data.remove("cursor")
-
-    def operations(self) -> list["Operation"]:
-        return [OperationChainWithDataExchange([Connect(), SelectOne(), Disconnect()])]
-
-
-class CreateAndDropTableWorkload(WorkloadWithContext, DdlWorkload):
-    def amend_data_before_execution(self, data: OperationData) -> None:
-        data.push("table_seed", data.get("worker_id"))
-
-    def operations(self) -> list["Operation"]:
-        return [
-            OperationChainWithDataExchange(
-                [CreateTableX(), PopulateTableX(), SelectStarFromTableX(), DropTableX()]
-            )
-        ]
-
-
-class CreateAndDropTableWithMvWorkload(WorkloadWithContext, DdlWorkload):
-    def amend_data_before_execution(self, data: OperationData) -> None:
-        data.push("table_seed", data.get("worker_id"))
-
-    def operations(self) -> list["Operation"]:
-        return [
-            OperationChainWithDataExchange(
-                [
-                    CreateTableX(),
-                    PopulateTableX(),
-                    CreateMvOnTableX(),
-                    SelectStarFromMvOnTableX(),
-                    DropTableX(),
-                ]
-            )
-        ]
