+++ misc/python/materialize/scalability/workloads/self_test_workloads.py
